dummyapp/requirements.txt
Flask==3.0.0
requests==2.31.0
PyYAML==6.0.1
python-dateutil==2.8.2
cryptography==41.0.7

